TwiAudioClassifier/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
PyYAML>=5.4.1
scikit-learn>=0.24.0
librosa>=0.9.0
augly>=0.2.0
